Flexible motor input

+++ emBODY/eBcode/arch-arm/embobj/plus/mc/AbsEncoder.c
@@ -117,7 +117,7 @@ void s_AbsEncoder_set_spikes_limis(AbsEncoder* o)
         case(eomc_enc_spichainof2):
         case(eomc_enc_spichainof3):
         {
-            int32_t toleranceIDeg = o->toleranceCfg * 65535 /360 ;
+            int32_t toleranceIDeg =(int32_t)((o->toleranceCfg * 65535.0f) / 360.0f) ;
             #define AEA_MIN_SPIKE 16 //iDegree 
             /* Note: AEA has 12 bits of resolution, so we pad with four zero to transform from AEA unit to iDegree */
             if(toleranceIDeg < AEA_MIN_SPIKE)
@@ -134,7 +134,7 @@ void s_AbsEncoder_set_spikes_limis(AbsEncoder* o)
         case(eomc_enc_psc):
         {
             
-            int32_t toleranceIDeg = o->toleranceCfg * 65535 /360 ;
+            int32_t toleranceIDeg = (int32_t)((o->toleranceCfg * 65535.0f) / 360.0f);
             o->spike_mag_limit = toleranceIDeg *o->div;
             o->spike_cnt_limit = AEA_DEFAULT_SPIKE_CNT_LIMIT; //ALE ??
             //snprintf(message, sizeof(message), "AMO: tol%.3f, div%.3f spikel%lu", o->toleranceCfg, o->div, o->spike_mag_limit);
